Removed models moved to hs-ontoloty-api; refactored routes to common_routes to escape name space conflicts
